backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tortoise-orm==0.20.0
aerich==0.7.2
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic[email]==2.5.0
pydantic-settings==2.1.0